requirements.txt
slackclient
psycopg2-binary
